+++ WeatherApp/WeatherApp/Controllers/UIExtensions.swift
@@ -1,0 +1,32 @@
+//
+//  UIExtensions.swift
+//  WeatherApp
+//
+//  Created by Phan Quyen on 18/08/2025.
+//
+
+import UIKit
+
+extension UIViewController {
+    
+    func showSuccessMessage(_ message: String) {
+        let alert = UIAlertController(
+            title: "Success",
+            message: message,
+            preferredStyle: .alert
+        )
+        alert.addAction(UIAlertAction(title: "OK", style: .default))
+        
+        present(alert, animated: true)
+        
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
+            alert.dismiss(animated: true)
+        }
+    }
+    
+    func showDetailAlert(title: String, message: String) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "OK", style: .default))
+        present(alert, animated: true)
+    }
+}
