+++ WeatherApp/WeatherApp/Core/AppDelegate.swift
@@ -12,7 +12,7 @@ import CoreData
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        debugPrint("App did finish launching")
+        //debugPrint("App did finish launching")
         setupInitialConfigurations()
         return true
     }
@@ -33,7 +33,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
     
     private func setupInitialConfigurations() {
-        debugPrint("Setting up initial configurations...")
+        //debugPrint("Setting up initial configurations...")
     }
 
     lazy var persistentContainer: NSPersistentContainer = {
